requirements.txt
streamlit
langchain-openai
langchain-community
langgraph
duckduckgo-search
python-dotenv
openai